@@ -1,13 +1,21 @@
 <?php
+/**
+ * This file is part of the login-cidadao project or it's bundles.
+ *
+ * (c) Guilherme Donato <guilhermednt on github>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace LoginCidadao\APIBundle\Controller;
 
 use FOS\RestBundle\Controller\FOSRestController;
 use JMS\Serializer\SerializationContext;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use LoginCidadao\CoreBundle\Entity\Authorization;
 use LoginCidadao\CoreBundle\Model\PersonInterface;
 use LoginCidadao\OAuthBundle\Model\ClientInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class BaseController extends FOSRestController
 {
@@ -23,6 +31,7 @@ class BaseController extends FOSRestController
 
         $view = $this->view($content)
             ->setSerializationContext($context);
+
         return $this->handleView($view);
     }
 
@@ -30,13 +39,15 @@ class BaseController extends FOSRestController
     {
         $person = $this->getUser();
         $serializer = $this->get('jms_serializer');
+
         return $serializer->serialize($person, 'json',
-                                        SerializationContext::create()->setGroups($scope));
+            SerializationContext::create()->setGroups($scope));
     }
 
-    protected function getClientScope(PersonInterface $user,
-                                        ClientInterface $client = null)
-    {
+    protected function getClientScope(
+        PersonInterface $user,
+        ClientInterface $client = null
+    ) {
         if ($client === null) {
             $client = $this->getClient();
         }
@@ -44,17 +55,18 @@ class BaseController extends FOSRestController
         $authorization = $this->getDoctrine()
             ->getRepository('LoginCidadaoCoreBundle:Authorization')
             ->findOneBy(array(
-            'person' => $user,
-            'client' => $client
-        ));
+                'person' => $user,
+                'client' => $client,
+            ));
         if (!($authorization instanceof Authorization)) {
-            throw new AccessDeniedHttpException("Access denied");
+            throw new AccessDeniedException("Access denied");
         }
 
         $scopes = $authorization->getScope();
         if (array_search('public', $scopes) === false) {
             $scopes[] = 'public';
         }
+
         return $scopes;
     }
 
@@ -72,9 +84,10 @@ class BaseController extends FOSRestController
     {
         $token = $this->get('security.token_storage')->getToken();
         $accessToken = $this->getDoctrine()->
-            getRepository('LoginCidadaoOAuthBundle:AccessToken')->
-            findOneBy(array('token' => $token->getToken()));
+        getRepository('LoginCidadaoOAuthBundle:AccessToken')->
+        findOneBy(array('token' => $token->getToken()));
         $client = $accessToken->getClient();
+
         return $client;
     }
 
