@@ -1,4 +1,12 @@
 <?php
+/**
+ * This file is part of the login-cidadao project or it's bundles.
+ *
+ * (c) Guilherme Donato <guilhermednt on github>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace LoginCidadao\CoreBundle\Controller;
 
@@ -6,12 +14,11 @@ use LoginCidadao\APIBundle\Entity\ActionLogRepository;
 use LoginCidadao\CoreBundle\Model\SupportMessage;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use LoginCidadao\CoreBundle\Entity\SentEmail;
 use LoginCidadao\APIBundle\Entity\LogoutKey;
 
@@ -38,38 +45,21 @@ class DefaultController extends Controller
         $form->handleRequest($request);
         $translator = $this->get('translator');
         $message = $translator->trans('contact.form.sent');
+
         if ($form->isValid()) {
-            $emailMessage = $data->getMessage();
-            if ($correlationId !== null) {
-                $emailMessage = "<p>$emailMessage</p><p>Correlation Id: {$correlationId}</p>";
-            }
-            $email = new SentEmail();
-            $email
-                ->setType('contact-mail')
-                ->setSubject('Fale conosco - '.$data->getName())
-                ->setSender($data->getEmail())
-                ->setReceiver($this->container->getParameter('contact_form.email'))
-                ->setMessage($emailMessage);
-            $mailer = $this->get('mailer');
+            $email = $this->getEmail($data, $correlationId);
             $swiftMail = $email->getSwiftMail();
-            if ($mailer->send($swiftMail)) {
+            if ($this->get('mailer')->send($swiftMail)) {
                 $em = $this->getDoctrine()->getManager();
                 $em->persist($email);
                 $em->flush();
                 $this->get('session')->getFlashBag()->add('success', $message);
             }
 
-            $url = $this->generateUrl("lc_contact");
-
-            return $this->redirect($url);
+            return $this->redirectToRoute('lc_contact');
         }
 
-        return $this->render(
-            'LoginCidadaoCoreBundle:Info:contact.html.twig',
-            array(
-                'form' => $form->createView(),
-            )
-        );
+        return $this->render('LoginCidadaoCoreBundle:Info:contact.html.twig', ['form' => $form->createView()]);
     }
 
     /**
@@ -112,7 +102,7 @@ class DefaultController extends Controller
         $logoutKey = $logoutKeys->findActiveByKey($key);
 
         if (!($logoutKey instanceof LogoutKey)) {
-            throw new AccessDeniedHttpException("Invalid logout key.");
+            throw new AccessDeniedException("Invalid logout key.");
         }
 
         $result['logged_out'] = false;
@@ -157,8 +147,23 @@ class DefaultController extends Controller
      */
     public function indexAction(Request $request, $lastUsername)
     {
-        return array(
-            'last_username' => $lastUsername,
-        );
+        ['last_username' => $lastUsername];
+    }
+
+    private function getEmail(SupportMessage $supportMessage, $correlationId = null)
+    {
+        $message = $supportMessage->getMessage();
+        if ($correlationId !== null) {
+            $message = "<p>$message</p><p>Correlation Id: {$correlationId}</p>";
+        }
+
+        $email = (new SentEmail())
+            ->setType('contact-mail')
+            ->setSubject('Fale conosco - '.$supportMessage->getName())
+            ->setSender($supportMessage->getEmail())
+            ->setReceiver($this->container->getParameter('contact_form.email'))
+            ->setMessage($message);
+
+        return $email;
     }
 }
