@@ -1,19 +1,24 @@
 <?php
+/**
+ * This file is part of the login-cidadao project or it's bundles.
+ *
+ * (c) Guilherme Donato <guilhermednt on github>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace LoginCidadao\CoreBundle\Controller;
 
+use LoginCidadao\CoreBundle\Security\TwoFactorAuthenticationService;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-use LoginCidadao\CoreBundle\Model\PersonInterface;
 use LoginCidadao\CoreBundle\Form\Type\TwoFactorAuthenticationFormType;
 use LoginCidadao\CoreBundle\Form\Type\TwoFactorAuthenticationDisableFormType;
 use LoginCidadao\CoreBundle\Form\Type\TwoFactorAuthenticationBackupCodeGenerationFormType;
 use Symfony\Component\Form\FormError;
-use LoginCidadao\CoreBundle\Entity\BackupCode;
-use Doctrine\ORM\EntityManager;
-use Scheb\TwoFactorBundle\Model\BackupCodeInterface;
 
 /**
  * @Route("/two-factor")
@@ -27,34 +32,33 @@ class TwoFactorAuthenticationController extends Controller
      */
     public function enableAction(Request $request)
     {
-        $twoFactor = $this->get("scheb_two_factor.security.google_authenticator");
+        /** @var TwoFactorAuthenticationService $twoFactor */
+        $twoFactor = $this->get('lc.two_factor');
+
         $translator = $this->get('translator');
-        $person = $this->getPerson();
-        $secret = $twoFactor->generateSecret();
-        $person->setGoogleAuthenticatorSecret($secret);
+        $person = $this->getUser();
+        $person->setGoogleAuthenticatorSecret($twoFactor->generateSecret());
 
         $form = $this->createForm(new TwoFactorAuthenticationFormType(), $person);
         $form->handleRequest($request);
 
         if ($form->isValid()) {
             $verificationCode = $form->get('verification')->getData();
-            $isValid = $twoFactor->checkCode($person, $verificationCode);
 
-            if ($isValid) {
-                $this->enable2FA($person, $form);
+            try {
+                $twoFactor->enable($form->getData(), $verificationCode);
+
                 $message = $translator->trans('Two-Factor Authentication enabled.');
                 $this->get('session')->getFlashBag()->add('success', $message);
+
                 return $this->redirect($this->generateUrl("fos_user_change_password"));
-            } else {
-                $message = $translator->trans('Invalid code! Make sure you configured your app correctly and your smartphone\'s time is adjusted.');
+            } catch (\InvalidArgumentException $e) {
+                $message = $translator->trans($e->getMessage());
                 $form->get('verification')->addError(new FormError($message));
             }
         }
 
-        return array(
-            'form' => $form->createView(),
-            'secretUrl' => $twoFactor->getUrl($person)
-        );
+        return ['form' => $form->createView(), 'secretUrl' => $twoFactor->getSecretUrl($person)];
     }
 
     /**
@@ -63,22 +67,23 @@ class TwoFactorAuthenticationController extends Controller
      */
     public function disableAction(Request $request)
     {
-        $person = $this->getPerson();
-        $form = $this->createForm(new TwoFactorAuthenticationDisableFormType(),
-                                    $person);
+        /** @var TwoFactorAuthenticationService $twoFactor */
+        $twoFactor = $this->get('lc.two_factor');
+
+        $person = $this->getUser();
+        $form = $this->createForm(new TwoFactorAuthenticationDisableFormType(), $person);
         $form->handleRequest($request);
 
         if ($form->isValid()) {
             $translator = $this->get('translator');
-            $this->disable2FA($person, $form);
+            $twoFactor->disable($person);
             $message = $translator->trans('Two-Factor Authentication disabled.');
             $this->get('session')->getFlashBag()->add('success', $message);
+
             return $this->redirect($this->generateUrl("fos_user_change_password"));
         }
 
-        return array(
-            'form' => $form->createView()
-        );
+        return ['form' => $form->createView()];
     }
 
     /**
@@ -87,25 +92,25 @@ class TwoFactorAuthenticationController extends Controller
      */
     public function generateBackupCodesAction(Request $request)
     {
-        $person = $this->getPerson();
-        $form = $this->createForm(new TwoFactorAuthenticationBackupCodeGenerationFormType(),
-                                    $person);
+        /** @var TwoFactorAuthenticationService $twoFactor */
+        $twoFactor = $this->get('lc.two_factor');
+
+        $person = $this->getUser();
+        $form = $this->createForm(new TwoFactorAuthenticationBackupCodeGenerationFormType(), $person);
         $form->handleRequest($request);
 
         if ($form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $translator = $this->get('translator');
-            $this->removeBackupCodes($em, $person);
-            $this->generateBackupCodes($em, $person);
-            $em->flush();
-            $message = $translator->trans('New Backup Codes generated. Don\'t forget to copy and store them safely.');
+            $twoFactor->removeBackupCodes($person);
+            $twoFactor->generateBackupCodes($person);
+
+            $message = $this->get('translator')
+                ->trans('New Backup Codes generated. Don\'t forget to copy and store them safely.');
             $this->get('session')->getFlashBag()->add('success', $message);
+
             return $this->redirect($this->generateUrl("fos_user_change_password"));
         }
 
-        return array(
-            'form' => $form->createView()
-        );
+        return ['form' => $form->createView()];
     }
 
     /**
@@ -114,58 +119,6 @@ class TwoFactorAuthenticationController extends Controller
      */
     public function formAction()
     {
-        return array();
+        return [];
     }
-
-    /**
-     * @return PersonInterface
-     */
-    protected function getPerson()
-    {
-        return $this->getUser();
-    }
-
-    protected function enable2FA(PersonInterface $person, $form)
-    {
-        $em = $this->getDoctrine()->getManager();
-        $this->generateBackupCodes($em, $person);
-        $em->persist($form->getData());
-        $em->flush();
-    }
-
-    protected function disable2FA(BackupCodeInterface $person, $form)
-    {
-        $em = $this->getDoctrine()->getManager();
-        $backupCodes = $person->getBackupCodes();
-        foreach ($backupCodes as $backupCode) {
-            $em->remove($backupCode);
-        }
-        $person->setGoogleAuthenticatorSecret(null);
-        $em->persist($person);
-        $em->flush();
-    }
-
-    protected function removeBackupCodes(EntityManager $em,
-                                            PersonInterface $person)
-    {
-        $backupCodes = $person->getBackupCodes();
-        foreach ($backupCodes as $backupCode) {
-            $em->remove($backupCode);
-        }
-    }
-
-    protected function generateBackupCodes(EntityManager $em,
-                                            PersonInterface $person)
-    {
-        $backupCodes = array();
-        while (count($backupCodes) < 10) {
-            $code = bin2hex(random_bytes(5));
-            $backupCode = new BackupCode();
-            $backupCode->setPerson($person);
-            $backupCode->setCode($code);
-            $backupCodes[] = $backupCode;
-            $em->persist($backupCode);
-        }
-    }
-
 }
