@@ -15,6 +15,7 @@ use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use libphonenumber\PhoneNumber;
+use LoginCidadao\CoreBundle\Tests\LongPolling\LongPollableInterface;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Validator\Constraints as Assert;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
@@ -25,7 +26,7 @@ use JMS\Serializer\Annotation as JMS;
 use FOS\UserBundle\Model\User as BaseUser;
 use LoginCidadao\OAuthBundle\Entity\Client;
 use LoginCidadao\CoreBundle\Model\LocationSelectData;
-use LoginCidadao\LongPolling\LongPollingUtils;
+use LoginCidadao\CoreBundle\LongPolling\LongPollingUtils;
 use LoginCidadao\CoreBundle\Model\PersonInterface;
 use LoginCidadao\OAuthBundle\Model\ClientInterface;
 use LoginCidadao\ValidationBundle\Validator\Constraints as LCAssert;
@@ -43,7 +44,7 @@ use Rollerworks\Bundle\PasswordStrengthBundle\Validator\Constraints as Rollerwor
  * @JMS\ExclusionPolicy("all")
  * @Vich\Uploadable
  */
-class Person extends BaseUser implements PersonInterface, TwoFactorInterface, BackupCodeInterface
+class Person extends BaseUser implements PersonInterface, BackupCodeInterface
 {
     /**
      * @ORM\Id
@@ -1223,49 +1224,6 @@ class Person extends BaseUser implements PersonInterface, TwoFactorInterface, Ba
     public function setPlaceOfBirth(LocationSelectData $location)
     {
         $location->toObject($this);
-    }
-
-    public function waitUpdate(EntityManagerInterface $em, \DateTime $updatedAt)
-    {
-        $id = $this->getId();
-        $lastUpdatedAt = null;
-        $callback = $this->getCheckUpdateCallback(
-            $em,
-            $id,
-            $updatedAt,
-            $lastUpdatedAt
-        );
-
-        return LongPollingUtils::runTimeLimited($callback);
-    }
-
-    private function getCheckUpdateCallback(
-        EntityManagerInterface $em,
-        $id,
-        $updatedAt,
-        $lastUpdatedAt
-    ) {
-        $people = $em->getRepository('LoginCidadaoCoreBundle:Person');
-
-        return function () use ($id, $people, $em, $updatedAt, $lastUpdatedAt) {
-            $em->clear();
-            $person = $people->find($id);
-            if (!$person->getUpdatedAt()) {
-                return false;
-            }
-
-            if ($person->getUpdatedAt() > $updatedAt) {
-                return $person;
-            }
-
-            if ($lastUpdatedAt === null) {
-                $lastUpdatedAt = $person->getUpdatedAt();
-            } elseif ($person->getUpdatedAt() != $lastUpdatedAt) {
-                return $person;
-            }
-
-            return false;
-        };
     }
 
     /**
