@@ -3,6 +3,9 @@
 namespace LoginCidadao\CoreBundle\Model;
 
 use Doctrine\ORM\EntityManagerInterface;
+use LoginCidadao\CoreBundle\Entity\BackupCode;
+use LoginCidadao\CoreBundle\Tests\LongPolling\LongPollableInterface;
+use Scheb\TwoFactorBundle\Model\Google\TwoFactorInterface;
 use Symfony\Component\HttpFoundation\File\File;
 use Doctrine\Common\Collections\ArrayCollection;
 use FOS\UserBundle\Model\UserInterface;
@@ -13,7 +16,7 @@ use Symfony\Component\Security\Core\Encoder\EncoderAwareInterface;
 use LoginCidadao\OAuthBundle\Entity\Client;
 use JMS\Serializer\Annotation as JMS;
 
-interface PersonInterface extends EncoderAwareInterface, UserInterface, LocationAwareInterface
+interface PersonInterface extends EncoderAwareInterface, UserInterface, LocationAwareInterface, LongPollableInterface, TwoFactorInterface
 {
     public function getId();
 
@@ -235,13 +238,6 @@ interface PersonInterface extends EncoderAwareInterface, UserInterface, Location
     public function setGoogleAuthenticatorSecret($googleAuthenticatorSecret);
 
     /**
-     * @param EntityManagerInterface $em
-     * @param \DateTime $updatedAt
-     * @return PersonInterface
-     */
-    public function waitUpdate(EntityManagerInterface $em, \DateTime $updatedAt);
-
-    /**
      * @return City
      */
     public function getCity();
@@ -271,4 +267,9 @@ interface PersonInterface extends EncoderAwareInterface, UserInterface, Location
      * @return ArrayCollection
      */
     public function getAddresses();
+
+    /**
+     * @return BackupCode[]
+     */
+    public function getBackupCodes();
 }
