@@ -1,4 +1,12 @@
 <?php
+/**
+ * This file is part of the login-cidadao project or it's bundles.
+ *
+ * (c) Guilherme Donato <guilhermednt on github>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace LoginCidadao\CoreBundle\Security\User\Manager;
 
@@ -6,24 +14,20 @@ use FOS\UserBundle\Doctrine\UserManager as BaseManager;
 use Doctrine\Common\Persistence\ObjectManager;
 use FOS\UserBundle\Util\CanonicalizerInterface;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
-use Symfony\Component\DependencyInjection\Container;
 use FOS\UserBundle\Model\UserInterface;
 use LoginCidadao\ValidationBundle\Validator\Constraints\UsernameValidator;
 
 class UserManager extends BaseManager
 {
 
-    private $container;
-
-    public function __construct(EncoderFactoryInterface $encoderFactory,
-                                CanonicalizerInterface $usernameCanonicalizer,
-                                CanonicalizerInterface $emailCanonicalizer,
-                                ObjectManager $om, $class, Container $container)
-    {
-        parent::__construct($encoderFactory, $usernameCanonicalizer,
-                $emailCanonicalizer, $om, $class);
-
-        $this->container = $container;
+    public function __construct(
+        EncoderFactoryInterface $encoderFactory,
+        CanonicalizerInterface $usernameCanonicalizer,
+        CanonicalizerInterface $emailCanonicalizer,
+        ObjectManager $om,
+        $class
+    ) {
+        parent::__construct($encoderFactory, $usernameCanonicalizer, $emailCanonicalizer, $om, $class);
     }
 
     public function createUser()
@@ -35,7 +39,7 @@ class UserManager extends BaseManager
      * Updates a user.
      *
      * @param UserInterface $user
-     * @param Boolean       $andFlush Whether to flush the changes (default true)
+     * @param Boolean $andFlush Whether to flush the changes (default true)
      */
     public function updateUser(UserInterface $user, $andFlush = true)
     {
@@ -52,7 +56,7 @@ class UserManager extends BaseManager
     public function enforceUsername(UserInterface $user)
     {
         $current = $user->getUsernameCanonical();
-        if (is_null($current) || strlen($current) == 0) {
+        if (is_null($current) || strlen($current) === 0) {
             $email = explode('@', $user->getEmailCanonical(), 2);
             $username = $email[0];
             if (!UsernameValidator::isUsernameValid($username)) {
@@ -71,11 +75,11 @@ class UserManager extends BaseManager
      * This is based on HWI's FOSUBRegistrationFormHandler
      *
      * @param string $username
-     * @param integer $username
+     * @param int $maxIterations
+     * @param string $default
      * @return string
      */
-    public function getNextAvailableUsername($username, $maxIterations = 10,
-                                                $default = null)
+    public function getNextAvailableUsername($username, $maxIterations = 10, $default = null)
     {
         $i = 0;
         $testName = $username;
@@ -84,24 +88,28 @@ class UserManager extends BaseManager
             $user = $this->findUserByUsername($testName);
         } while ($user !== null && $i < $maxIterations && $testName = $username.$i++);
 
-        if (is_null($user)) {
+        if (!$user) {
             return $testName;
         } else {
-            if (is_null($default)) {
-                return "$username@".time();
-            } else {
-                return $default;
-            }
+            return $default ?: "$username@".time();
         }
     }
 
-    public function findUserByUsernameOrEmail($username) {
-        if (is_numeric($cpf = preg_replace('/[^0-9]/', '', $username)) && strlen($cpf) == 11) {
-            $person = parent::findUserBy(array('cpf' => $cpf));
+    public function findUserByUsernameOrEmail($username)
+    {
+        $onlyNumbers = preg_replace('/[^0-9]/', '', $username);
+        $looksLikeCpf = is_numeric($onlyNumbers) && strlen($onlyNumbers) == 11;
+
+        // If it looks like a CPF we give it a try
+        if ($looksLikeCpf) {
+            $person = parent::findUserBy(['cpf' => $onlyNumbers]);
+
             if ($person !== null) {
                 return $person;
             }
-        } 
-        return parent::findUserByUsernameOrEmail($username);        
+        }
+
+        // If it doesn't look like a CPF number or if we couldn't find nobody with that CPF, fallback to this:
+        return parent::findUserByUsernameOrEmail($username);
     }
 }
