@@ -2,7 +2,6 @@
 
 namespace LoginCidadao\OpenIDBundle\Controller;
 
-use Doctrine\ORM\EntityManager;
 use LoginCidadao\OAuthBundle\Entity\Client;
 use JMS\Serializer\SerializationContext;
 use Symfony\Component\HttpFoundation\Request;
@@ -29,8 +28,7 @@ class ClientRegistrationController extends FOSRestController
         $form->handleRequest($request);
         if ($form->isValid()) {
             $metadata = $form->getData();
-            $em = $this->getDoctrine()->getManager();
-            $client = $this->registerClient($em, $metadata);
+            $client = $this->registerClient($metadata);
 
             return $this->view($metadata->fromClient($client), 201);
         } else {
@@ -101,8 +99,9 @@ class ClientRegistrationController extends FOSRestController
      * @param ClientMetadata $data
      * @return Client
      */
-    private function registerClient(EntityManager $em, ClientMetadata $data)
+    private function registerClient(ClientMetadata $data)
     {
+        $em = $this->getDoctrine()->getManager();
         if ($data->getClient() === null) {
             $client = $data->toClient();
         } else {
